Split impl
